refactor(agent): keep ACP client within size limit
